cleaned the code of productcontroller

+++ app/Http/Controllers/ProductController.php
@@ -2,11 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreProductRequest;
+use App\Http\Requests\UpdateProductRequest;
 use App\Models\Category;
 use App\Models\Product;
+use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Psy\CodeCleaner\AssignThisVariablePass;
 
 class ProductController extends Controller
 {
@@ -18,9 +22,10 @@ class ProductController extends Controller
     public function index()
     {
         $products = Product::all();
+
         return response()->json([
-            'status' => 'success',
-            'products' => $products,
+            '========' => '================== Display Products : ==================',
+            'Products' => $products,
         ]);
     }
 
@@ -30,16 +35,14 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\jsonResponse
      */
-    public function store(Request $request)
+    public function store(StoreProductRequest $request)
     {
         $user = Auth::user();
-
         $product = Product::create($request->all() + ['user_id' => $user->id]);
 
         return response()->json([
-            'status' => true,
-            'message' => 'Product added successfully!',
-            'product' => $product,
+            'Message' => 'Product added successfully!',
+            'Product' => $product,
         ], 201);
     }
 
@@ -52,10 +55,8 @@ class ProductController extends Controller
     public function show($id)
     {
         $product = Product::find($id);
-        if(!$product){
-            return response()->json(['message' => 'Sorry, this product doesn\'t exist!',]);
-        }
-        return response()->json($product, 200);
+        $response = ($product) ? response()->json($product, 200) : response()->json(['message' => "Product with (id : {$id}) doesn't exist!",]);
+        return $response;
     }
 
     /**
@@ -65,19 +66,22 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\jsonResponse
      */
-    public function update(Request $request, $id)
+    public function update(UpdateProductRequest $request, $id)
     {
         $user = Auth::user();
         $product = Product::find($id);
 
+        if(!$product){
+            return response()->json(['message' => "Product with (id : {$id} doesn't exist!)"]);
+        }
         if(!$user->can('edit every product') && $user->id != $product->user_id){
-            return response()->json(['message' => 'Sorry, this is not your product!']);
+            return response()->json(['message' => "Can't update a product that isn't yours!"]);
         }
 
         $product->update($request->all());
+
         return response()->json([
-            'status' => true,
-            'message' => 'Product updated successfully!',
+            '*******' => '**************** Product updated successfully! ****************',
             'product' => $product,
         ], 200);
     }
@@ -93,11 +97,15 @@ class ProductController extends Controller
         $user = Auth::user();
         $product = Product::find($id);
 
-        if(!$user->can('delete every product') && $user->id != $product->user_id){
-            return response()->json(['message' => 'Sorry, this is not your product!']);
+        if(!$product){
+            return response()->json(['message' => "Product with (id : {$id} doesn't exist!)"]);
+        }
+        if (!$user->can('delete every product') && $user->id != $product->user_id){
+            return response()->json(['message' => "Product with this (id : {$id} doesn't exist!)"], 403);
         }
 
         $product->delete();
+
         return response()->json([
             'status' => true,
             'message' => 'Product delete successfully!',
@@ -114,8 +122,5 @@ class ProductController extends Controller
 //        return response()->json([
 //            'product' => $products,
 //        ]);
-    }
-    public function test(){
-        return 'test';
     }
 }
